+++ stowPackages/neovim/.config/nvim/lua/plugins/config/barbar.lua
@@ -11,8 +11,8 @@ vim.g.bufferline = {
   no_name_title = "Empty file", -- empty buffer title
 
   -- Configure icons on the bufferline.
-  icon_separator_active   = '', -- '',  -- '▎',
-  icon_separator_inactive = '', -- '',  -- '▎',
+  icon_separator_active   = '', -- '', -- '',  -- '▎',
+  icon_separator_inactive =	'', -- '', -- '', -- '',  -- '▎',
   icon_close_tab          = '',
   icon_close_tab_modified = '●',
   icon_pinned             = '車',