+++ stowPackages/neovim/.config/nvim/lua/plugins/config/lualine.lua
@@ -8,11 +8,11 @@ end
 lualine.setup {
   options = {
     icons_enabled = true,
-    theme = 'auto',
-    component_separators = { left = '', right = ''},
-    section_separators = { left = '', right = ''},
-    --section_separators = { left = '', right = '' },
-    --component_separators = { left = '', right = '' },
+    theme = 'palenight',
+    --component_separators = { left = '', right = ''},
+    --section_separators = { left = '', right = ''},
+    section_separators = { left = '', right = '' },
+    component_separators = { left = '', right = '' },
     disabled_filetypes = {},
     always_divide_middle = true,
     globalstatus = true,